fix encoding content error with search in forums (only php 5.5.6?)
when click on search link in a forum page or with search form in sidebar
display an encoding error in web browser.
search.php?search_within[]=forums === error encoding content blank page.
I think after search in code, It's an oversights of removing [] in links
search_within with [] = page error, with not = page ok

Error found with this config
PHP: 5.5.6-1
MYSQL: 5.5.33-1